Second Commit
In this second commit, I tried to start learning CMake from the tutorial on their WebSite. I found it too hard, and decided to try the OpenCV+CMake Tutorial on OpenCV website. It was great, simple and I've already got something working. Now I'm motivated to keep on the hard job :)